img2pdf/pyproject.toml
[tool.poetry]
name = "img2pdf"
version = "0.1.0"
description = ""
authors = ["Nina Jensen <jnik@aarhus.dk>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
